Combined both 3D components into new Gh3DLossesGainsVisualizer component
